proof of possession factory, further code cleanup
